refactor: change ListSection, LoadingSkeleton, NoResults, SearchBar, and SearchHistoryDropdow to be reusable components to avoid duplication
