--- src/modules/TourSearchResultsModule/module-elements/LoadingSkeleton.tsx
@@ -1,28 +1,0 @@
-import React from 'react'
-import { Skeleton } from '@/components/ui/skeleton'
-import { Card, CardContent } from '@/components/ui/card'
-import { v4 } from 'uuid'
-
-interface LoadingSkeletonProps {
-  count?: number
-}
-
-const LoadingSkeleton: React.FC<LoadingSkeletonProps> = ({ count = 6 }) => {
-  return (
-    <>
-      {Array.from({ length: count }).map((_, index) => (
-        <Card key={`${index}-${v4()}`} className="overflow-hidden">
-          <Skeleton className="h-24 sm:h-40 w-full" />
-          <CardContent className="p-3 space-y-2">
-            <Skeleton className="h-4 w-full" />
-            <Skeleton className="h-3 w-2/3" />
-            <Skeleton className="h-3 w-3/4" />
-            <Skeleton className="h-4 w-3/4" />
-          </CardContent>
-        </Card>
-      ))}
-    </>
-  )
-}
-
-export default LoadingSkeleton
